requirements.txt
flask
psycopg2
pandas
plotly
reportlab
python-dotenv
gunicorn